requirements.txt
PyQt5
opencv-python
imutils
SpeechRecognition==2.1.3
numpy
pyserial
tqdm
pillow
TextToGcode
keyboard
vosk
pyaudio
gcode2image